+++ sentry-android-core/src/main/java/io/sentry/android/core/performance/AppStartMetrics.java
@@ -1,0 +1,208 @@
+package io.sentry.android.core.performance;
+
+import android.app.Application;
+import android.content.ContentProvider;
+import android.os.SystemClock;
+import androidx.annotation.Nullable;
+import io.sentry.android.core.ContextUtils;
+import io.sentry.android.core.SentryAndroidOptions;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * An in-memory representation for app-metrics during app start. As the SDK can't be initialized
+ * that early, we can't use transactions or spans directly. Thus simple TimeSpans are used and later
+ * transformed into SDK specific txn/span data structures.
+ */
+@ApiStatus.Internal
+public class AppStartMetrics {
+
+  public enum AppStartType {
+    UNKNOWN,
+    COLD,
+    WARM
+  }
+
+  private static volatile @Nullable AppStartMetrics instance;
+
+  private @NotNull AppStartType appStartType = AppStartType.UNKNOWN;
+  private boolean appLaunchedInForeground = false;
+
+  private final @NotNull TimeSpan appStartSpan;
+  private final @NotNull TimeSpan sdkInitTimeSpan;
+  private final @NotNull TimeSpan applicationOnCreate;
+  private final @NotNull Map<ContentProvider, TimeSpan> contentProviderOnCreates;
+  private final @NotNull List<ActivityLifecycleTimeSpan> activityLifecycles;
+
+  public static @NotNull AppStartMetrics getInstance() {
+
+    if (instance == null) {
+      synchronized (AppStartMetrics.class) {
+        if (instance == null) {
+          instance = new AppStartMetrics();
+        }
+      }
+    }
+    //noinspection DataFlowIssue
+    return instance;
+  }
+
+  public AppStartMetrics() {
+    appStartSpan = new TimeSpan();
+    sdkInitTimeSpan = new TimeSpan();
+    applicationOnCreate = new TimeSpan();
+    contentProviderOnCreates = new HashMap<>();
+    activityLifecycles = new ArrayList<>();
+  }
+
+  /**
+   * @return the app start span Uses Process.getStartUptimeMillis() as start timestamp, which
+   *     requires API level 24+
+   */
+  public @NotNull TimeSpan getAppStartTimeSpan() {
+    return appStartSpan;
+  }
+
+  /**
+   * @return the SDK init time span, as measured pre-performance-v2 Uses ContentProvider/Sdk init
+   *     time as start timestamp
+   *     <p>Data is filled by either {@link io.sentry.android.core.SentryPerformanceProvider} with a
+   *     fallback to {@link io.sentry.android.core.SentryAndroid}. At least the start timestamp
+   *     should always be set.
+   */
+  public @NotNull TimeSpan getSdkInitTimeSpan() {
+    return sdkInitTimeSpan;
+  }
+
+  public @NotNull TimeSpan getApplicationOnCreateTimeSpan() {
+    return applicationOnCreate;
+  }
+
+  public void setAppStartType(final @NotNull AppStartType appStartType) {
+    this.appStartType = appStartType;
+  }
+
+  public @NotNull AppStartType getAppStartType() {
+    return appStartType;
+  }
+
+  public boolean isAppLaunchedInForeground() {
+    return appLaunchedInForeground;
+  }
+
+  /**
+   * Provides all collected content provider onCreate time spans
+   *
+   * @return A sorted list of all onCreate calls
+   */
+  public @NotNull List<TimeSpan> getContentProviderOnCreateTimeSpans() {
+    final List<TimeSpan> measurements = new ArrayList<>(contentProviderOnCreates.values());
+    Collections.sort(measurements);
+    return measurements;
+  }
+
+  public @NotNull List<ActivityLifecycleTimeSpan> getActivityLifecycleTimeSpans() {
+    final List<ActivityLifecycleTimeSpan> measurements = new ArrayList<>(activityLifecycles);
+    Collections.sort(measurements);
+    return measurements;
+  }
+
+  public void addActivityLifecycleTimeSpans(final @NotNull ActivityLifecycleTimeSpan timeSpan) {
+    activityLifecycles.add(timeSpan);
+  }
+
+  /**
+   * @return the app start time span if it was started and perf-2 is enabled, falls back to the sdk
+   *     init time span otherwise
+   */
+  public @NotNull TimeSpan getAppStartTimeSpanWithFallback(
+      final @NotNull SentryAndroidOptions options) {
+    if (options.isEnablePerformanceV2()) {
+      // Only started when sdk version is >= N
+      final @NotNull TimeSpan appStartSpan = getAppStartTimeSpan();
+      if (appStartSpan.hasStarted()) {
+        return appStartSpan;
+      }
+    }
+
+    // fallback: use sdk init time span, as it will always have a start time set
+    return getSdkInitTimeSpan();
+  }
+
+  public void clear() {
+    appStartType = AppStartType.UNKNOWN;
+    appStartSpan.reset();
+    sdkInitTimeSpan.reset();
+    applicationOnCreate.reset();
+    contentProviderOnCreates.clear();
+    activityLifecycles.clear();
+  }
+
+  /**
+   * Called by instrumentation
+   *
+   * @param application The application object where onCreate was called on
+   * @noinspection unused
+   */
+  public static void onApplicationCreate(final @NotNull Application application) {
+    final long now = SystemClock.uptimeMillis();
+
+    final @NotNull AppStartMetrics instance = getInstance();
+    if (instance.applicationOnCreate.hasNotStarted()) {
+      instance.applicationOnCreate.setStartedAt(now);
+      instance.appLaunchedInForeground = ContextUtils.isForegroundImportance();
+    }
+  }
+
+  /**
+   * Called by instrumentation
+   *
+   * @param application The application object where onCreate was called on
+   * @noinspection unused
+   */
+  public static void onApplicationPostCreate(final @NotNull Application application) {
+    final long now = SystemClock.uptimeMillis();
+
+    final @NotNull AppStartMetrics instance = getInstance();
+    if (instance.applicationOnCreate.hasNotStopped()) {
+      instance.applicationOnCreate.setDescription(application.getClass().getName() + ".onCreate");
+      instance.applicationOnCreate.setStoppedAt(now);
+    }
+  }
+
+  /**
+   * Called by instrumentation
+   *
+   * @param contentProvider The content provider where onCreate was called on
+   * @noinspection unused
+   */
+  public static void onContentProviderCreate(final @NotNull ContentProvider contentProvider) {
+    final long now = SystemClock.uptimeMillis();
+
+    final TimeSpan measurement = new TimeSpan();
+    measurement.setStartedAt(now);
+    getInstance().contentProviderOnCreates.put(contentProvider, measurement);
+  }
+
+  /**
+   * Called by instrumentation
+   *
+   * @param contentProvider The content provider where onCreate was called on
+   * @noinspection unused
+   */
+  public static void onContentProviderPostCreate(final @NotNull ContentProvider contentProvider) {
+    final long now = SystemClock.uptimeMillis();
+
+    final @Nullable TimeSpan measurement =
+        getInstance().contentProviderOnCreates.get(contentProvider);
+    if (measurement != null && measurement.hasNotStopped()) {
+      measurement.setDescription(contentProvider.getClass().getName() + ".onCreate");
+      measurement.setStoppedAt(now);
+    }
+  }
+}