+++ sentry-android-core/src/test/java/io/sentry/android/core/ActivityLifecycleIntegrationTest.kt
@@ -32,6 +32,8 @@ import io.sentry.TraceContext
 import io.sentry.TransactionContext
 import io.sentry.TransactionFinishedCallback
 import io.sentry.TransactionOptions
+import io.sentry.android.core.performance.AppStartMetrics
+import io.sentry.android.core.performance.AppStartMetrics.AppStartType
 import io.sentry.protocol.MeasurementValue
 import io.sentry.protocol.TransactionNameSource
 import io.sentry.test.DeferredExecutorService
@@ -76,7 +78,6 @@ class ActivityLifecycleIntegrationTest {
             dsn = "https://key@sentry.io/proj"
         }
         val bundle = mock<Bundle>()
-        val context = TransactionContext("name", "op")
         val activityFramesTracker = mock<ActivityFramesTracker>()
         val fullyDisplayedReporter = FullyDisplayedReporter.getInstance()
         val transactionFinishedCallback = mock<TransactionFinishedCallback>()
@@ -96,9 +97,10 @@ class ActivityLifecycleIntegrationTest {
             whenever(hub.options).thenReturn(options)
 
             // We let the ActivityLifecycleIntegration create the proper transaction here
-            val argumentCaptor = argumentCaptor<TransactionOptions>()
-            whenever(hub.startTransaction(any(), argumentCaptor.capture())).thenAnswer {
-                val t = SentryTracer(context, hub, argumentCaptor.lastValue)
+            val optionCaptor = argumentCaptor<TransactionOptions>()
+            val contextCaptor = argumentCaptor<TransactionContext>()
+            whenever(hub.startTransaction(contextCaptor.capture(), optionCaptor.capture())).thenAnswer {
+                val t = SentryTracer(contextCaptor.lastValue, hub, optionCaptor.lastValue)
                 transaction = t
                 return@thenAnswer t
             }
@@ -130,7 +132,7 @@ class ActivityLifecycleIntegrationTest {
 
     @BeforeTest
     fun `reset instance`() {
-        AppStartState.getInstance().resetInstance()
+        AppStartMetrics.getInstance().clear()
 
         context = ApplicationProvider.getApplicationContext()
         val activityManager = context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager?
@@ -610,7 +612,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityDestroyed(activity)
 
         val span = fixture.transaction.children.first()
-        assertEquals(span.status, SpanStatus.CANCELLED)
+        assertEquals(SpanStatus.CANCELLED, span.status)
         assertTrue(span.isFinished)
     }
 
@@ -795,7 +797,7 @@ class ActivityLifecycleIntegrationTest {
         val activity = mock<Activity>()
         sut.onActivityCreated(activity, null)
 
-        assertTrue(AppStartState.getInstance().isColdStart!!)
+        assertEquals(AppStartType.COLD, AppStartMetrics.getInstance().appStartType)
     }
 
     @Test
@@ -808,7 +810,7 @@ class ActivityLifecycleIntegrationTest {
         val bundle = Bundle()
         sut.onActivityCreated(activity, bundle)
 
-        assertFalse(AppStartState.getInstance().isColdStart!!)
+        assertEquals(AppStartType.WARM, AppStartMetrics.getInstance().appStartType)
     }
 
     @Test
@@ -822,7 +824,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityCreated(activity, bundle)
         sut.onActivityCreated(activity, null)
 
-        assertFalse(AppStartState.getInstance().isColdStart!!)
+        assertEquals(AppStartType.WARM, AppStartMetrics.getInstance().appStartType)
     }
 
     @Test
@@ -831,14 +833,19 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        val date = SentryNanotimeDate(Date(0), 0)
+        val date = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(date)
 
         val activity = mock<Activity>()
         sut.onActivityCreated(activity, fixture.bundle)
 
         // call only once
-        verify(fixture.hub).startTransaction(any(), check<TransactionOptions> { assertEquals(date, it.startTimestamp) })
+        verify(fixture.hub).startTransaction(
+            any(),
+            check<TransactionOptions> {
+                assertEquals(date.nanoTimestamp(), it.startTimestamp!!.nanoTimestamp())
+            }
+        )
     }
 
     @Test
@@ -848,9 +855,9 @@ class ActivityLifecycleIntegrationTest {
         sut.register(fixture.hub, fixture.options)
 
         // usually set by SentryPerformanceProvider
-        val date = SentryNanotimeDate(Date(0), 0)
+        val date = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(date)
-        AppStartState.getInstance().setAppStartEnd(1)
+        AppStartMetrics.getInstance().sdkInitTimeSpan.setStoppedAt(2)
 
         val activity = mock<Activity>()
         sut.onActivityCreated(activity, fixture.bundle)
@@ -866,12 +873,13 @@ class ActivityLifecycleIntegrationTest {
         sut.register(fixture.hub, fixture.options)
 
         // usually set by SentryPerformanceProvider
-        val startDate = SentryNanotimeDate(Date(0), 0)
+        val startDate = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(startDate)
-        AppStartState.getInstance().setColdStart(false)
-        AppStartState.getInstance().setAppStartEnd(1)
+        val appStartMetrics = AppStartMetrics.getInstance()
+        appStartMetrics.appStartType = AppStartType.WARM
+        appStartMetrics.sdkInitTimeSpan.setStoppedAt(2)
 
-        val endDate = AppStartState.getInstance().appStartEndTime!!
+        val endDate = appStartMetrics.sdkInitTimeSpan.projectedStopTimestamp
 
         val activity = mock<Activity>()
         sut.onActivityCreated(activity, fixture.bundle)
@@ -879,7 +887,7 @@ class ActivityLifecycleIntegrationTest {
         val appStartSpanCount = fixture.transaction.children.count {
             it.spanContext.operation.startsWith("app.start.warm") &&
                 it.startDate.nanoTimestamp() == startDate.nanoTimestamp() &&
-                it.finishDate!!.nanoTimestamp() == endDate.nanoTimestamp()
+                it.finishDate!!.nanoTimestamp() == endDate!!.nanoTimestamp()
         }
         assertEquals(1, appStartSpanCount)
     }
@@ -892,9 +900,9 @@ class ActivityLifecycleIntegrationTest {
 
         // usually done by SentryPerformanceProvider, if disabled it's done by
         // SentryAndroid.init
-        val startDate = SentryNanotimeDate(Date(0), 0)
+        val startDate = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(startDate)
-        AppStartState.getInstance().setColdStart(false)
+        AppStartMetrics.getInstance().appStartType = AppStartType.WARM
 
         // when activity is created
         val view = fixture.createView()
@@ -902,14 +910,14 @@ class ActivityLifecycleIntegrationTest {
         whenever(activity.findViewById<View>(any())).thenReturn(view)
         sut.onActivityCreated(activity, fixture.bundle)
         // then app-start end time should still be null
-        assertNull(AppStartState.getInstance().appStartEndTime)
+        assertTrue(AppStartMetrics.getInstance().sdkInitTimeSpan.hasNotStopped())
 
         // when activity is resumed
         sut.onActivityResumed(activity)
         Thread.sleep(1)
         runFirstDraw(view)
         // end-time should be set
-        assertNotNull(AppStartState.getInstance().appStartEndTime)
+        assertTrue(AppStartMetrics.getInstance().sdkInitTimeSpan.hasStopped())
     }
 
     @Test
@@ -919,10 +927,10 @@ class ActivityLifecycleIntegrationTest {
         sut.register(fixture.hub, fixture.options)
 
         // usually done by SentryPerformanceProvider
-        val startDate = SentryNanotimeDate(Date(0), 0)
+        val startDate = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(startDate)
-        AppStartState.getInstance().setColdStart(false)
-        AppStartState.getInstance().setAppStartEnd(1234)
+        AppStartMetrics.getInstance().appStartType = AppStartType.WARM
+        AppStartMetrics.getInstance().sdkInitTimeSpan.setStoppedAt(1234)
 
         // when activity is created and resumed
         val activity = mock<Activity>()
@@ -932,7 +940,7 @@ class ActivityLifecycleIntegrationTest {
         // then the end time should not be overwritten
         assertEquals(
             DateUtils.millisToNanos(1234),
-            AppStartState.getInstance().appStartEndTime!!.nanoTimestamp()
+            AppStartMetrics.getInstance().sdkInitTimeSpan.projectedStopTimestamp!!.nanoTimestamp()
         )
     }
 
@@ -943,9 +951,9 @@ class ActivityLifecycleIntegrationTest {
         sut.register(fixture.hub, fixture.options)
 
         // usually done by SentryPerformanceProvider
-        val startDate = SentryNanotimeDate(Date(0), 0)
+        val startDate = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(startDate)
-        AppStartState.getInstance().setColdStart(false)
+        AppStartMetrics.getInstance().appStartType = AppStartType.WARM
 
         // when activity is created, started and resumed multiple times
         val view = fixture.createView()
@@ -957,7 +965,7 @@ class ActivityLifecycleIntegrationTest {
         Thread.sleep(1)
         runFirstDraw(view)
 
-        val firstAppStartEndTime = AppStartState.getInstance().appStartEndTime
+        val firstAppStartEndTime = AppStartMetrics.getInstance().sdkInitTimeSpan.projectedStopTimestamp
 
         Thread.sleep(1)
         sut.onActivityPaused(activity)
@@ -970,7 +978,7 @@ class ActivityLifecycleIntegrationTest {
         // then the end time should not be overwritten
         assertEquals(
             firstAppStartEndTime!!.nanoTimestamp(),
-            AppStartState.getInstance().appStartEndTime!!.nanoTimestamp()
+            AppStartMetrics.getInstance().sdkInitTimeSpan.projectedStopTimestamp!!.nanoTimestamp()
         )
     }
 
@@ -980,7 +988,7 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        val date = SentryNanotimeDate(Date(0), 0)
+        val date = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(date)
 
         val activity = mock<Activity>()
@@ -988,7 +996,7 @@ class ActivityLifecycleIntegrationTest {
 
         val span = fixture.transaction.children.first()
         assertEquals(span.operation, "app.start.warm")
-        assertSame(span.startDate, date)
+        assertEquals(span.startDate.nanoTimestamp(), date.nanoTimestamp())
     }
 
     @Test
@@ -997,7 +1005,7 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        val date = SentryNanotimeDate(Date(0), 0)
+        val date = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(date)
 
         val activity = mock<Activity>()
@@ -1005,7 +1013,7 @@ class ActivityLifecycleIntegrationTest {
 
         val span = fixture.transaction.children.first()
         assertEquals(span.operation, "app.start.cold")
-        assertSame(span.startDate, date)
+        assertEquals(span.startDate.nanoTimestamp(), date.nanoTimestamp())
     }
 
     @Test
@@ -1014,7 +1022,7 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        val date = SentryNanotimeDate(Date(0), 0)
+        val date = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(date)
 
         val activity = mock<Activity>()
@@ -1022,7 +1030,7 @@ class ActivityLifecycleIntegrationTest {
 
         val span = fixture.transaction.children.first()
         assertEquals(span.description, "Warm Start")
-        assertSame(span.startDate, date)
+        assertEquals(span.startDate.nanoTimestamp(), date.nanoTimestamp())
     }
 
     @Test
@@ -1031,7 +1039,7 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        val date = SentryNanotimeDate(Date(0), 0)
+        val date = SentryNanotimeDate(Date(1), 0)
         setAppStartTime(date)
 
         val activity = mock<Activity>()
@@ -1039,7 +1047,7 @@ class ActivityLifecycleIntegrationTest {
 
         val span = fixture.transaction.children.first()
         assertEquals(span.description, "Cold Start")
-        assertSame(span.startDate, date)
+        assertEquals(span.startDate.nanoTimestamp(), date.nanoTimestamp())
     }
 
     @Test
@@ -1048,7 +1056,7 @@ class ActivityLifecycleIntegrationTest {
         fixture.options.tracesSampleRate = 1.0
         sut.register(fixture.hub, fixture.options)
 
-        val date = SentryNanotimeDate(Date(0), 0)
+        val date = SentryNanotimeDate(Date(1), 0)
         setAppStartTime()
 
         val activity = mock<Activity>()
@@ -1486,8 +1494,18 @@ class ActivityLifecycleIntegrationTest {
         shadowOf(Looper.getMainLooper()).idle()
     }
 
-    private fun setAppStartTime(date: SentryDate = SentryNanotimeDate(Date(0), 0)) {
+    private fun setAppStartTime(date: SentryDate = SentryNanotimeDate(Date(1), 0)) {
         // set by SentryPerformanceProvider so forcing it here
-        AppStartState.getInstance().setAppStartTime(0, date)
+        val sdkAppStartTimeSpan = AppStartMetrics.getInstance().sdkInitTimeSpan
+        val appStartTimeSpan = AppStartMetrics.getInstance().appStartTimeSpan
+        val millis = DateUtils.nanosToMillis(date.nanoTimestamp().toDouble()).toLong()
+
+        sdkAppStartTimeSpan.setStartedAt(millis)
+        sdkAppStartTimeSpan.setStartUnixTimeMs(millis)
+        sdkAppStartTimeSpan.setStoppedAt(0)
+
+        appStartTimeSpan.setStartedAt(millis)
+        appStartTimeSpan.setStartUnixTimeMs(millis)
+        appStartTimeSpan.setStoppedAt(0)
     }
 }